Move sort component into seperate file
